Atualizar main com develop - oficina-os-service v1.0.15

+++ src/main/java/br/com/oficina/os/framework/messaging/OutboxPublisher.java
@@ -86,7 +86,7 @@ public class OutboxPublisher {
     }
 
     static long retryMultiplier(int attempts) {
-        var retryIndex = Math.clamp((long) attempts - 1L, 0L, 10L);
+        var retryIndex = Math.clamp(attempts - 1L, 0L, 10L);
         return 1L << retryIndex;
     }
